requirements.txt
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
